pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nanobot-todoist-skill"
version = "1.0.0"
description = "Todoist MCP server for Nanobot - create todos when nanobot reminds you"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "your.email@example.com" }]
keywords = ["nanobot", "mcp", "todoist", "todo", "reminder"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "todoist-api-python>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
todoist-mcp = "todoist_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/steeveroucaute10-epping/nanobot-todoist-skill"
Repository = "https://github.com/steeveroucaute10-epping/nanobot-todoist-skill"

[tool.setuptools.packages.find]
where = ["src"]
